refactor: 기존 클래스 사용처 제거, Legacy 접두사 붙임
