+++ Assets/Scripts/Card/InStage/HandBar/HandBarModel.cs
@@ -9,39 +9,39 @@ namespace Cardevil.Card.InStage
     [Serializable]
     public class HandBarModel
     {
-        [SerializeReference] private List<INewCardState> hand = new();
-        [SerializeReference] private List<INewCardState> selection = new();
+        [SerializeReference] private List<ICardState> hand = new();
+        [SerializeReference] private List<ICardState> selection = new();
 
         [field: Space] 
         [field: SerializeField] public InteractionData PointerDownData { get; private set; }
         [field: SerializeField] public InteractionData DragData { get; private set; }
         [field: SerializeField] public InteractionData DetachData { get; private set; }
         
-        private readonly HashSet<INewCardState> handSet = new();
-        private readonly HashSet<INewCardState> selectionSet = new();
+        private readonly HashSet<ICardState> handSet = new();
+        private readonly HashSet<ICardState> selectionSet = new();
 
-        public IReadOnlyList<INewCardState> Hand => hand;
-        public IReadOnlyList<INewCardState> Selection => selection;
+        public IReadOnlyList<ICardState> Hand => hand;
+        public IReadOnlyList<ICardState> Selection => selection;
 
         /// <summary>
         /// 손패 상에 위치한 순서대로 정렬된 선택 목록.
         /// 호출 시 새로운 객체를 반환함.
         /// </summary>
-        public IReadOnlyList<INewCardState> SortedSelection => selection.OrderBy(s => hand.IndexOf(s)).ToList();
+        public IReadOnlyList<ICardState> SortedSelection => selection.OrderBy(s => hand.IndexOf(s)).ToList();
 
-        public void Add(INewCardState state)
+        public void Add(ICardState state)
         {
             if (!handSet.Add(state)) return;
             hand.Add(state);
         }
 
-        public void Insert(int index, INewCardState state)
+        public void Insert(int index, ICardState state)
         {
             if (!handSet.Add(state)) return;
             hand.Insert(index, state);
         }
 
-        public void Remove(INewCardState state)
+        public void Remove(ICardState state)
         {
             if (!handSet.Remove(state)) return;
             hand.Remove(state);
@@ -49,13 +49,13 @@ namespace Cardevil.Card.InStage
                 selection.Remove(state);
         }
 
-        public void Select(INewCardState state)
+        public void Select(ICardState state)
         {
             if (selectionSet.Add(state))
                 selection.Add(state);
         }
 
-        public void Deselect(INewCardState state)
+        public void Deselect(ICardState state)
         {
             if (selectionSet.Remove(state))
                 selection.Remove(state);
@@ -65,7 +65,7 @@ namespace Cardevil.Card.InStage
         /// 카드를 들고 HandBar에서 뺄 때 호출함.
         /// 정보를 저장하고 리스트에서 제거함.
         /// </summary>
-        public void Detach(INewCardState state)
+        public void Detach(ICardState state)
         {
             Remove(state);
 
@@ -76,7 +76,7 @@ namespace Cardevil.Card.InStage
         /// HandBar에서 뺀(들고 있는) 카드를 다시 넣을 때 호출함.
         /// 정보를 바탕으로 리스트에 복구함.
         /// </summary>
-        public void Reattach(INewCardState state)
+        public void Reattach(ICardState state)
         {
             Debug.Assert(state.Id == DetachData.Card.Id);
 
@@ -88,14 +88,14 @@ namespace Cardevil.Card.InStage
             ClearDetachData();
         }
 
-        public int IndexOf(INewCardState state) => hand.IndexOf(state);
+        public int IndexOf(ICardState state) => hand.IndexOf(state);
 
         public void Swap(int indexA, int indexB)
         {
             (hand[indexA], hand[indexB]) = (hand[indexB], hand[indexA]);
         }
 
-        public void Sort(IComparer<INewCardState> comparer)
+        public void Sort(IComparer<ICardState> comparer)
         {
             hand.Sort(comparer.Compare);
         }
@@ -103,12 +103,12 @@ namespace Cardevil.Card.InStage
         [Serializable]
         public struct InteractionData
         {
-            [field: SerializeReference] public INewCardState Card { get; private set; }
+            [field: SerializeReference] public ICardState Card { get; private set; }
             [field: SerializeField] public int OriginalIndex { get; private set; }
             [field: SerializeField] public bool IsSelected { get; private set; }
             [field: SerializeField] public float LastInteractionTime { get; private set; }
 
-            public InteractionData(INewCardState card, int originalIndex, bool isSelected)
+            public InteractionData(ICardState card, int originalIndex, bool isSelected)
             {
                 Card = card;
                 OriginalIndex = originalIndex;
@@ -121,7 +121,7 @@ namespace Cardevil.Card.InStage
             public static InteractionData Empty => new(null, -1, false);
         }
         
-        public void SetPointerDownData(INewCardState state)
+        public void SetPointerDownData(ICardState state)
         {
             PointerDownData = new InteractionData(state, IndexOf(state), selectionSet.Contains(state));
         }
@@ -131,7 +131,7 @@ namespace Cardevil.Card.InStage
             PointerDownData = InteractionData.Empty;
         }
 
-        public void SetDraggingData(INewCardState state)
+        public void SetDraggingData(ICardState state)
         {
             DragData = new InteractionData(state, IndexOf(state), selectionSet.Contains(state));
         }
@@ -141,7 +141,7 @@ namespace Cardevil.Card.InStage
             DragData = InteractionData.Empty;
         }
 
-        public void SetDetachData(INewCardState state)
+        public void SetDetachData(ICardState state)
         {
             DetachData = new InteractionData(state, IndexOf(state), selectionSet.Contains(state));
         }
