+++ Assets/Scripts/Card/_Common/_Core/CardState.cs.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: 9c190b3dcd964c8087fea68d0fbfd196
-timeCreated: 1770623226
+guid: 918467d2e02f4025918f7e6c2b648669
+timeCreated: 1778302045
