+++ Assets/Scripts/Card/_Common/_Core/CardStateBuilder.cs.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: fd1f18169aaa439e8a80824f8f8e909e
-timeCreated: 1770626812
+guid: b507a8006941499eb900193235de1c34
+timeCreated: 1778301387
