+++ Assets/Scripts/Card/_Common/_Core/Elements/BaseColorElement.cs
@@ -18,13 +18,8 @@ namespace Cardevil.Card.Common.Core
         {
             return new BaseColorElement(color);
         }
-        
-        public void Apply(CardStateBuilder builder)
-        {
-            builder.SetDefaultColor(color);
-        }
 
-        public void Apply(NewCardStateBuilder builder)
+        public void Apply(CardStateBuilder builder)
         {
             builder.DefaultColor = color;
         }
