+++ Assets/Scripts/Card/_Common/_Core/Legacy.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 21157388c40548abbf8fe2bfd19086ea
+timeCreated: 1778333563
