+++ Assets/Scripts/Card/_Common/_Core/Legacy/LegacyCardState.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 9c190b3dcd964c8087fea68d0fbfd196
+timeCreated: 1770623226
