+++ Assets/Scripts/Card/_Common/_Core/Legacy/LegacyCardStateBuilder.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: fd1f18169aaa439e8a80824f8f8e909e
+timeCreated: 1770626812
