--- Assets/Scripts/Card/_Common/_Core/NewCardState.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 918467d2e02f4025918f7e6c2b648669
-timeCreated: 1778302045
