--- Assets/Scripts/Card/_Common/_Core/NewCardStateBuilder.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: b507a8006941499eb900193235de1c34
-timeCreated: 1778301387
